requirements-dev.txt
tensorflow==1.11.0
numpy==1.15.3
Pillow==5.3.0
pytest==3.9.3